refactor adding/modifying/showing pet implementation, fix errors, improve imports

+++ src/views/EditPetView.js
@@ -1,15 +1,26 @@
 import React, { useEffect, useState } from 'react';
 import { connect, useDispatch } from 'react-redux';
 import { makeStyles } from '@material-ui/core/styles';
-import { PetInformationForm, PetPhotosForm } from 'components/forms';
 import { Button, CircularProgress } from '@material-ui/core';
-import { usePet, useProfilePicture } from 'helper/hooks/pets.hooks';
-import { changePet, clearPetInfos, getPet, updateProfilePicture, updateSelectedPet } from 'redux/actions';
-import { useUser } from 'helper/hooks/auth.hooks';
-import Loading from 'components/Loading';
-import { NotificationService, FirebaseService } from 'services';
 import { useHistory, useLocation } from 'react-router-dom';
+// components import
+import { PetInformationForm, PetPhotosForm } from 'components/forms';
+import Loading from 'components/Loading';
+// state imports
+import { usePet, usePetCompetitions, usePetDocuments, usePetPictures, usePetProfilePictureToRemove, usePetProfilePictureToUpload } from 'helper/hooks/pets.hooks';
+import { changePet, getPet, updateCompetitionsToUpload, updateDocumentsToUpload, updatePicturesToUpload, updateProfilePicture } from 'redux/actions';
+// helpers import
+import { useUser } from 'helper/hooks/auth.hooks';
+import { UPLOAD_STATUS } from 'helper/types';
+import { isObjEmpty } from 'helper/helpers';
+// services import
+import { NotificationService, FirebaseService } from 'services';
 
+/**
+ * 
+ * @param {*} props 
+ * Main view to edit/modify an existing pet
+ */
 const EditPetView = (props) => {
     const classes = useStyles();
     const dispatch = useDispatch();
@@ -20,9 +31,32 @@ const EditPetView = (props) => {
     const pet = usePet();
     const user = useUser();
 
+    // get pet upload states
+    const petDocuments = usePetDocuments();
+    const petCompetitions = usePetCompetitions();
+    const petPictures = usePetPictures();
+    const petProfilePictureToUpload = usePetProfilePictureToUpload();
+    const petProfilePictureToRemove = usePetProfilePictureToRemove();
+
+    const [loading, setLoading] = useState(true);
+    const [buttonLoading, setButtonLoading] = useState(false);
+    const [formIsDisabled, setFormIsDisabled] = useState(false);
+    const [name, setName] = useState(pet.officialName);
+    const [nickname, setNickname] = useState(pet.nickname);
+    const [sex, setSex] = useState(pet.sex);
+    const [birthDate, setBirthDate] = useState(new Date(pet.birthDate));
+    const [species, setSpecies] = useState(pet.species);
+    const [breed, setBreed] = useState(pet.breed);
+    const [price, setPrice] = useState(pet.price);
+
     useEffect(() => {
-        // remove old profile picture
-        dispatch(updateProfilePicture({}));
+        // update upload states
+        if (loading) {
+            dispatch(updateProfilePicture(pet.profilePicture, {}));
+            dispatch(updatePicturesToUpload(pet.pictures));
+            dispatch(updateCompetitionsToUpload(pet.competitions));
+            dispatch(updateDocumentsToUpload(pet.documents));
+        }
 
         if (pet.officialName) {
             setLoading(false);
@@ -33,24 +67,14 @@ const EditPetView = (props) => {
             };
             fetchPet();
         }
-    }, [dispatch, id, pet.officialName]);
-
-    const [loading, setLoading] = useState(true);
-    const [formIsDisabled, setFormIsDisabled] = useState(false);
-    const [name, setName] = useState(pet.officialName);
-    const [nickname, setNickname] = useState(pet.nickname);
-    const [sex, setSex] = useState(pet.sex);
-    const [birthDate, setBirthDate] = useState(new Date(pet.birthDate));
-    const [species, setSpecies] = useState(pet.species);
-    const [breed, setBreed] = useState(pet.breed);
-    const [price, setPrice] = useState(pet.price);
+    }, [dispatch, id, pet.documents, pet.competitions, pet.profilePicture, pet.pictures, pet.officialName, loading]);
 
     const isEmpty = (str) => str === '' || str === undefined;
 
     useEffect(() => {
-        const disabled = isEmpty(name) || isEmpty(sex) || isEmpty(species) || isEmpty(breed);
+        const disabled = isEmpty(name) || isObjEmpty(petProfilePictureToUpload) || isEmpty(sex) || isEmpty(species) || isEmpty(breed);
         setFormIsDisabled(disabled);
-    }, [name, sex, breed, species]);
+    }, [name, sex, breed, species, petProfilePictureToUpload]);
 
     useEffect(() => {
         if (pet.ownerId !== user.id) {
@@ -59,82 +83,82 @@ const EditPetView = (props) => {
         }
     }, [pet.ownerId, user.id, history]);
 
-    // get old profile picture to delete it later if it was updated
-    const oldProfilePicture = useProfilePicture();
-
     const uploadCompetitions = async () => {
-        const competitionsData = [...pet.competitions];
+        const competitionsData = [...petCompetitions];
         for (let index = 0; index < competitionsData.length; index++) {
             let value = competitionsData[index];
 
-            if (value.certificate) {
-                if (value.certificate.status === 'upload') {
+            if (value.certificate && Object.keys(value.certificate).length !== 0) {
+                if (value.certificate.status === UPLOAD_STATUS.UPLOAD) {
                     const metadata = {
                         contentType: value.certificate.type,
                     };
                     let url = await FirebaseService.upload(value.certificate.path, value.certificate.data, metadata);
                     competitionsData[index].certificate.url = url;
-                } else if (value.certificate.status === 'delete') {
+                } else if (value.certificate.status === UPLOAD_STATUS.DELETE) {
                     await FirebaseService.remove(value.certificate.path);
                 }
             }
         }
 
-        let petData = pet;
-        petData.competitions = competitionsData.filter((value) => value.certificate.status !== 'delete');
-        await dispatch(updateSelectedPet(petData));
+        let competitions = competitionsData.filter((value) => value.certificate.status !== UPLOAD_STATUS.DELETE);
+        await dispatch(updateCompetitionsToUpload(competitions));
+
+        return competitions;
     };
 
     const uploadDocuments = async () => {
-        const documentsData = [...pet.documents];
+        const documentsData = [...petDocuments];
         for (let index = 0; index < documentsData.length; index++) {
             let value = documentsData[index];
 
-            if (value.status === 'upload') {
+            if (value.status === UPLOAD_STATUS.UPLOAD) {
                 const metadata = {
                     contentType: value.type,
                 };
                 let url = await FirebaseService.upload(value.path, value.data, metadata);
                 documentsData[index].url = url;
-            } else if (value.status === 'delete') {
+            } else if (value.status === UPLOAD_STATUS.DELETE) {
                 await FirebaseService.remove(value.path);
             }
         }
 
-        let petData = pet;
-        petData.documents = documentsData.filter((value) => value.status !== 'delete');
-        await dispatch(updateSelectedPet(petData));
+        let docs = documentsData.filter((value) => value.status !== UPLOAD_STATUS.DELETE);
+        await dispatch(updateDocumentsToUpload(docs));
+
+        return docs;
     };
 
     const uploadPictures = async () => {
-        const picturesData = [...pet.pictures];
+        const picturesData = [...petPictures];
         for (let index = 0; index < picturesData.length; index++) {
             let value = picturesData[index];
 
-            if (value.status === 'upload') {
+            if (value.status === UPLOAD_STATUS.UPLOAD) {
                 const metadata = {
                     contentType: 'image/png',
                 };
 
                 let url = await FirebaseService.upload(value.path, value.data, metadata);
                 picturesData[index].src = url;
-            } else if (value.status === 'delete') {
+            } else if (value.status === UPLOAD_STATUS.DELETE) {
                 await FirebaseService.remove(value.path);
             }
         }
-        let petData = pet;
-        petData.pictures = picturesData.filter((value) => value.status !== 'delete');
-        await dispatch(updateSelectedPet(petData));
+        const pics = picturesData.filter((value) => value.status !== UPLOAD_STATUS.DELETE);
+        await dispatch(updatePicturesToUpload(pics));
+
+        return pics;
     };
 
     const uploadProfilePicture = async () => {
         // update profile picture only if new was uploaded
-        if (oldProfilePicture.src && oldProfilePicture.src !== '') {
+        if (Object.keys(petProfilePictureToRemove).length !== 0 && petProfilePictureToRemove?.src !== '') {
             // delete old one
-            await FirebaseService.remove(oldProfilePicture.path);
+            await FirebaseService.remove(petProfilePictureToRemove.path);
 
             // upload new one
-            let value = pet.profilePicture;
+            let value = petProfilePictureToUpload;
             const metadata = {
                 contentType: 'image/png',
             };
@@ -142,56 +166,53 @@ const EditPetView = (props) => {
             let url = await FirebaseService.upload(value.path, value.data, metadata);
             value.src = url;
 
-            let petData = pet;
-            petData.profilePicture = value;
+            await dispatch(updateProfilePicture(value, {}));
 
-            await dispatch(updateSelectedPet(petData));
+            return value;
+        } else {
+            return pet.profilePicture;
         }
     };
 
     const updatePet = async () => {
-        if (pet.profilePicture.path) {
-            setLoading(true);
+        setButtonLoading(true);
 
-            // upload documents and pics to firebase first
-            await uploadDocuments();
-            await uploadCompetitions();
-            await uploadPictures();
-            await uploadProfilePicture();
+        // upload documents and pics to firebase first
+        const documents = await uploadDocuments();
+        const competitions = await uploadCompetitions();
+        const pictures = await uploadPictures();
+        const profilePicture = await uploadProfilePicture();
 
-            const dateCreated = Date.now();
-            // combine all information about a pet
-            let petToUpload = {
-                id: id,
-                ownerId: user.id,
-                officialName: name,
-                nickname: nickname,
-                birthDate: birthDate,
-                sex: sex,
-                price: price,
-                profilePicture: pet.profilePicture,
-                pictures: pet.pictures,
-                dateCreated: dateCreated,
-                breed: breed,
-                species: species,
-                competitions: pet.competitions,
-                documents: pet.documents,
-            };
+        const dateCreated = Date.now();
+        // combine all information about a pet
+        let petToUpload = {
+            id: id,
+            ownerId: user.id,
+            officialName: name,
+            nickname: nickname,
+            birthDate: birthDate,
+            sex: sex,
+            price: price,
+            profilePicture: profilePicture,
+            pictures: pictures,
+            dateCreated: dateCreated,
+            breed: breed,
+            species: species,
+            competitions: competitions,
+            documents: documents,
+        };
 
-            const onSuccess = () => {
-                NotificationService.notify('success', 'Success', 'Your four-legged friend was successfully updated!');
-                history.push('/');
-            };
+        const onSuccess = () => {
+            NotificationService.notify('success', 'Success', 'Your four-legged friend was successfully updated!');
+            history.push('/');
+        };
 
-            const onError = () => {
-                NotificationService.notify('error', 'Error', 'There was a problem updating your pet.');
-            };
+        const onError = () => {
+            NotificationService.notify('error', 'Error', 'There was a problem updating your pet.');
+        };
 
-            dispatch(changePet(petToUpload, onSuccess, onError));
-            setLoading(false);
-        } else {
-            NotificationService.notify('error', 'Upload Error', 'Upload of the pet profile picture is required!');
-        }
+        await dispatch(changePet(petToUpload, onSuccess, onError));
+        setButtonLoading(false);
     };
 
     // on canceling the view
@@ -218,7 +239,7 @@ const EditPetView = (props) => {
                     Cancel
                 </Button>
                 <Button disabled={formIsDisabled} onClick={updatePet} type="submit" variant="contained" color="primary" size="large">
-                    {loading ? <CircularProgress size={20} color="white" style={{ marginRight: 10 }} /> : ''} Save
+                    {buttonLoading ? <CircularProgress size={20} style={{ marginRight: 10, color: 'white' }} /> : ''} Save
                 </Button>
             </div>
         </div>
